+++ constants.h
@@ -11,6 +11,6 @@ enum types_event
 const int WIDTH  = 1920;
 const int HEIGHT = 1080;
 
-const double SCROLL_COEFF = 0.1;
+const float SCROLL_COEFF = 0.1;
 
 const char *TIMES_NEW_ROMAN_FONT = "fonts/Times_New_Roman.ttf";
